Wait until `data-final-fixes` to select tech and recipe
Previously, whether or not this mod would pick up other mods would
depend on the loading order, which is not well-defined by Factorio.

This change updates the mod to wait until the last loading phase
to choose its unlock technology and recipe ingredients, which
should make it more robust at interoperating with other mods when
the load order changes.